skills/github-app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
cryptography==41.0.7
pytest==7.4.3
pytest-cov==4.1.0
